+++ src/components/Table/TR/index.js
@@ -4,6 +4,7 @@ import classNames from 'classnames';
 const TR = (props) => {
   let classes = classNames({
     'oui-table-row--active': props.isActive,
+    'no-border': props.noBorder,
   });
   return (
     <tr
@@ -19,6 +20,8 @@ TR.propTypes = {
   children: React.PropTypes.node,
   /** If true, add active class */
   isActive: React.PropTypes.bool,
+  /** If true, add class to remove border */
+  noBorder: React.PropTypes.bool,
   /** Hook for automated JavaScript tests */
   testSection: React.PropTypes.string,
 };